requirements_pip.txt
python-chess>=0.30.0